Move withValueReplacement into src/HOCs/
